+++ Server/Logic.cpp
@@ -138,7 +138,7 @@ namespace GObject
 
     void Logic::SendMsg(int sessionID,int cmdid,const void * buffer,int size)
     {
-        Stream st(static_cast<UInt16>(cmdid),
+        Stream st(static_cast<UInt32>(cmdid),
             (static_cast<UInt16>(1) << 8) + static_cast<UInt16>(1), // target 
             (static_cast<UInt16>(0xFF) << 8) + static_cast<UInt16>(1) // source
             );
